refactor mockups for cleaner structure
